pyproject.toml
[tool.poetry]
name = "erer-dashboard"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "erer_dashboard"}]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.91.0"
alembic = "^1.9.3"
psycopg2-binary = "^2.9.5"
uvicorn = "^0.20.0"
python-dotenv = "^0.21.1"
fastapi-sqlalchemy = "^0.2.1"
pydantic = "^1.10.4"


[tool.poetry.group.dev.dependencies]
autopep8 = "^2.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
